remove util module -> refactor client&pathtype
