+++ core/src/pathtype.rs
