--- core/src/util/mod.rs
@@ -1,3 +1,0 @@
-pub use pathtype::PathType;
-
-mod pathtype;
